Refactor scaling of area attributes
- ScaledAttributes -> AreaAttributesScaler
- Compute scaling_factor in top-level class Scaler
